+++ haskell/app/Api.hs
@@ -1,7 +1,0 @@
-module Main (main) where
-
-import qualified CLI
-import Prelude (IO)
-
-main :: IO ()
-main = CLI.mainApi