+++ haskell/src/CLI.hs
@@ -3,40 +3,80 @@
 {-# OPTIONS_GHC -Wno-partial-fields #-}
 
 -- | The CLI entrpoints
-module CLI (mainApi, mainMacroscope, mainLentille) where
+module CLI (main) where
 
-import Env
+import Env hiding (Parser, auto, footer)
+import qualified Env
 import qualified Lentille.Gerrit as G
 import Macroscope.Main (runMacroscope)
 import qualified Monocle.Api
 import Monocle.Client (withClient)
 import Monocle.Prelude hiding ((:::))
 import Monocle.Search.Query (parseDateValue)
-import Options.Generic
+import Options.Applicative hiding (header, help, str)
+import qualified Options.Applicative as O
+import Options.Applicative.Help.Pretty (string)
 import qualified Streaming.Prelude as S
 
 ---------------------------------------------------------------
--- API cli
+-- Unified CLI
 ---------------------------------------------------------------
 
-data CliEnv = CliEnv
-  { config :: String,
-    elastic_conn :: String
-  }
+-- | usage is an optparse-applicative Parser that provides the final action
+-- See the last example of https://github.com/pcapriotti/optparse-applicative#commands
+usage :: Parser (IO ())
+usage =
+  subparser
+    ( mkCommand "Start the API" "api" usageApi (Just usageApiEnv)
+        <> mkCommand "Start the Crawlers" "crawler" usageCrawler (Just usageCrawlerEnv)
+        <> mkCommand "Maintain the database" "janitor" usageJanitor Nothing
+        <> mkCommand "Run a single crawler standlone" "lentille" usageLentille Nothing
+    )
+  where
+    -- The API entrypoint (no CLI argument).
+    usageApi :: O.Parser (IO ())
+    usageApi = pure $ do
+      -- get parameters from the environment
+      (config, elastic, port) <- getFromEnv usageApiEnv
+      -- start the API
+      Monocle.Api.run (getInt port) (getURL elastic) config
+    usageApiEnv :: Env.Parser Env.Error (FilePath, String, String)
+    usageApiEnv = (,,) <$> envConf <*> envElastic <*> envApiPort
 
-cliEnv :: IO CliEnv
-cliEnv =
-  Env.parse (header "monocle-api available environement variables") $
-    CliEnv <$> var str "CONFIG" (help "The Monocle configuration" <> def "/etc/monocle/config.yaml" <> helpDef show)
-      <*> var str "ELASTIC_CONN" (help "The Elasticsearch endpoint" <> def "elastic:9200")
+    -- The Crawler entrypoint (no CLI argument).
+    usageCrawler = pure $ do
+      -- get parameters from the environment
+      (config, url, monitoringPort) <- getFromEnv usageCrawlerEnv
+      -- start the Crawler
+      withClient url Nothing $ runMacroscope (getInt monitoringPort) config
+    usageCrawlerEnv = (,,) <$> envConf <*> envPublicUrl <*> envMonitoring
 
-newtype CLIOptions w = CLIOption
-  { port :: w ::: Maybe Int <?> "The API port to listen to, default to 9898"
-  }
-  deriving stock (Generic)
+    -- The janitor entrypoint
+    usageJanitor = pure $ putStrLn "NotImplemented"
 
-instance ParseRecord (CLIOptions Wrapped) where
-  parseRecord = parseRecordWithModifiers lispCaseModifiers
+    -- Helper to create sub command
+    mkEnvDoc envParser = string (Env.helpDoc envParser)
+    mkCommand doc name parser envParser = command name $ info (parser <**> helper) (progDesc doc <> extraHelp)
+      where
+        -- We only add `--help` to sub command which uses environment
+        extraHelp = maybe idm (footerDoc . Just . mkEnvDoc) envParser
+
+    -- Helpers to get value fromm the env
+    getFromEnv = Env.parse (header "monocle")
+    envConf = var str "CONFIG" (help "The Monocle configuration" <> def "/etc/monocle/config.yaml" <> helpDef show)
+    envElastic = var str "ELASTIC_CONN" (help "The Elasticsearch endpoint" <> def "elastic:9200")
+    envApiPort = var str "MONOCLE_API_PORT" (help "The API Port" <> def "9898")
+    envPublicUrl = var str "MONOCLE_PUBLIC_URL" (help "The Monocle URL" <> def "http://web:8080")
+    envMonitoring = var str "MONOCLE_CRAWLER_MONITORING" (help "The Monitoring Port" <> def "9001")
+    getInt txt = fromMaybe (error . from $ "Invalid number: " <> txt) $ readMaybe txt
+
+main :: IO ()
+main = join $ execParser opts
+  where
+    opts =
+      info
+        (usage <**> helper)
+        (fullDesc <> progDesc "changemetrics.io | monocle")
 
 getURL :: String -> Text
 getURL url =
@@ -44,51 +84,53 @@ getURL url =
       url' = if hasScheme then url else "http://" <> url
    in toText url'
 
-mainApi :: IO ()
-mainApi = do
-  -- Fetch environement variables
-  CliEnv {config, elastic_conn} <- cliEnv
-  -- Run arguments parser
-  CLIOption port' <- unwrapRecord "Monocle API"
-  -- Run the Monocle API
-  Monocle.Api.run (fromMaybe 8989 port') (getURL elastic_conn) config
-
----------------------------------------------------------------
--- Macroscope cli
----------------------------------------------------------------
-
-data Macroscope w = Macroscope
-  { monocleUrl :: w ::: Maybe Text <?> "The monocle API",
-    port :: w ::: Int <!> "9001" <?> "Health check port",
-    config :: w ::: Maybe FilePath <?> "The monocle configuration"
-  }
-  deriving stock (Generic)
-
-instance ParseRecord (Macroscope Wrapped) where
-  parseRecord = parseRecordWithModifiers lispCaseModifiers
-
-mainMacroscope :: IO ()
-mainMacroscope = do
-  Macroscope monocleUrl' port' config'' <- unwrapRecord "Macroscope lentille runner"
-  config' <- fromMaybe (error "--config or CONFIG env is required") <$> lookupEnv "CONFIG"
-  withClient (fromMaybe "http://web:8080" monocleUrl') Nothing $ \client ->
-    runMacroscope
-      port'
-      (fromMaybe config' config'')
-      client
-
 ---------------------------------------------------------------
 -- Lentille cli
 ---------------------------------------------------------------
+usageLentille :: Parser (IO ())
+usageLentille =
+  subparser
+    ( mkCommand "gerrit-change" "Get a single change" gerritChangeUsage
+        <> mkCommand "gerrit-projects" "Get projects list" gerritProjectsUsage
+        <> mkCommand "gerrit-changes" "Get changes list" gerritChangesUsage
+    )
+  where
+    gerritChangeUsage = io <$> parser
+      where
+        parser = (,) <$> urlOption <*> changeOption
+        io (url, change) = do
+          env <- getGerritEnv url
+          dump Nothing $ G.streamChange env [G.ChangeId $ show (change :: Int)]
 
-data Lentille
-  = GerritChange {url :: Text, change :: Int}
-  | GerritProjects {url :: Text, query :: Text}
-  | GerritChanges {url :: Text, project :: Text, since :: Text, limit :: Maybe Int}
-  deriving stock (Generic)
+    gerritProjectsUsage = io <$> parser
+      where
+        parser = (,) <$> urlOption <*> queryOption
+        io (url, query) = do
+          env <- getGerritEnv url
+          dump Nothing $ G.streamProject env $ G.Regexp query
 
-instance ParseRecord Lentille where
-  parseRecord = parseRecordWithModifiers lispCaseModifiers
+    gerritChangesUsage = io <$> parser
+      where
+        parser = (,,,) <$> urlOption <*> projectOption <*> sinceOption <*> limitOption
+        io (url, project, since, limit) = do
+          env <- getGerritEnv url
+          dump limit $ G.streamChange env [G.Project project, G.After (toSince since)]
+
+    toSince txt = case Monocle.Search.Query.parseDateValue txt of
+      Just x -> x
+      Nothing -> error $ "Invalid date: " <> show txt
+    getGerritEnv url = do
+      client <- G.getGerritClient url Nothing
+      pure $ G.GerritEnv client Nothing (const Nothing) "cli"
+
+    urlOption = strOption (long "url" <> O.help "Gerrit API url" <> metavar "URL")
+    queryOption = strOption (long "query" <> O.help "Gerrit regexp query")
+    changeOption = option auto (long "change" <> O.help "Change Number" <> metavar "NR")
+    projectOption = strOption (long "project" <> O.help "Project name")
+    sinceOption = strOption (long "since" <> O.help "Since date")
+    limitOption = optional $ option auto (long "limit" <> O.help "Limit count")
+
+    mkCommand name doc parser = command name $ info (parser <**> helper) $ progDesc doc
 
 dump :: (MonadCatch m, MonadIO m, ToJSON a) => Maybe Int -> Stream (Of a) m () -> m ()
 dump limitM stream = do
@@ -99,24 +141,3 @@ dump limitM stream = do
   liftIO . putLBSLn =<< exportMetricsAsText
   where
     brk = maybe id S.take limitM
-
-mainLentille :: IO ()
-mainLentille = do
-  args <- getRecord "Lentille runner"
-  case args of
-    GerritChange url change -> do
-      env <- getGerritEnv url
-      dump Nothing $ G.streamChange env [G.ChangeId $ show change]
-    GerritProjects url query -> do
-      env <- getGerritEnv url
-      dump Nothing $ G.streamProject env $ G.Regexp query
-    GerritChanges url project since limit -> do
-      env <- getGerritEnv url
-      dump limit $ G.streamChange env [G.Project project, G.After (toSince since)]
-  where
-    toSince txt = case Monocle.Search.Query.parseDateValue txt of
-      Just x -> x
-      Nothing -> error $ "Invalid date: " <> show txt
-    getGerritEnv url = do
-      client <- G.getGerritClient url Nothing
-      pure $ G.GerritEnv client Nothing (const Nothing) "cli"